Cleanup FunctionalNodeModule
